backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clear-sdk-python",
    "fastapi",
    "openai",
    "python-dotenv",
    "sqlalchemy",
    "uvicorn",
    "PyPDF2",
    "python-docx",
    "google-generativeai",
]
